feat: Synthesise shard lag offset when possible
Also incluse splitting large `replicore_cluster_view` files into smaller units.
