+++ core/cluster/view/src/lib.rs
@@ -2,5 +2,6 @@ mod error;
 mod view;
 
 pub use self::error::ClusterViewCorrupt;
+pub use self::error::ManyPrimariesFound;
 pub use self::view::ClusterView;
 pub use self::view::ClusterViewBuilder;
