+++ core/util/errors/Cargo.toml
@@ -1,0 +1,16 @@
+[package]
+name = "replicore_util_errors"
+version = "0.1.0"
+authors = ["Stefano Pogliani <stefano@spogliani.net>"]
+edition = "2021"
+
+description = "Utilites to better support errors"
+documentation = "https://www.replicante.io/docs"
+homepage = "https://www.replicante.io/"
+repository = "https://github.com/replicante-io/replicante"
+license = "MIT"
+
+
+[dependencies]
+anyhow = "^1.0.53"
+failure = "^0.1.5"
